+++ changelog.md
@@ -1,5 +1,5 @@
 **BugFixes**
-- KeyBinds doesn't work when screen is opened
+- KeyBinds doesn't work when screen is opened. With that change when a key is pressed now the notification is sent on the screen and not in the chat
 
 **Internal**
 - Improve Modrinth publishing